перене useitem, добавил места покупок, и пару возможностей в бомжа

+++ HoboConsolePrjct/Model/Items/Base/IItem.cs
@@ -12,7 +12,7 @@ namespace HoboConsole.Model.Items.Base
 {
     public interface IItem : IEntity, IEffect
     {
-        public decimal Price { get; }
+        public int Price { get; }
         public string Name { get; }
         public int Pleasure { get; } //Определяет как и как сильно влияет купленная вещь на эмоц. состояние
     }
