+++ HoboConsolePrjct/Model/Items/Drugs.cs
@@ -15,7 +15,7 @@ namespace HoboConsole.Model.Items
     public class Drugs : IItem, IEntity
     {
         public Guid Id { get; set; }
-        public decimal Price { get; }
+        public int Price { get; }
         public string Name { get; }
         public int Pleasure { get; } //Определяет как и как сильно влияет купленная вещь на эмоц. состояние
         public int Healthy { get; } //Определяет насколько увеличится или уменьшится здоровье
@@ -23,7 +23,7 @@ namespace HoboConsole.Model.Items
         [JsonConverter(typeof(StringEnumConverter))]
         public ItemTypeEnum ItemType { get; }
 
-        public Drugs(Guid id, decimal price, string name, int pleasure, int healthy, int energyBoost, ItemTypeEnum itemType)
+        public Drugs(Guid id, int price, string name, int pleasure, int healthy, int energyBoost, ItemTypeEnum itemType)
         {
             Id = id;
             Price = price;
