+++ HoboConsolePrjct/Model/Items/Garbage.cs
@@ -15,14 +15,14 @@ namespace HoboConsole.Model.Items
     public class Garbage : IItem, IEntity
     {
         public Guid Id { get; set; }
-        public decimal Price { get; }
+        public int Price { get; }
         public string Name { get; }
         public int Pleasure { get; }
         public int Healthy { get; }
         [JsonConverter(typeof(StringEnumConverter))]
         public ItemTypeEnum ItemType { get; }
 
-        public Garbage(Guid id, decimal price, string name, int pleasure, int healthy, ItemTypeEnum itemTypeEnum)
+        public Garbage(Guid id, int price, string name, int pleasure, int healthy, ItemTypeEnum itemTypeEnum)
         {
             Id = id;
             Price = price;
